backend/requirements.txt
asgiref==3.8.1
astroid==3.3.8
bandit==1.8.2
cfgv==3.4.0
colorama==0.4.6
coverage==7.6.10
dill==0.3.9
distlib==0.3.9
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.16.1
identify==2.6.5
isort==5.13.2
mando==0.7.1
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
nodeenv==1.9.1
pbr==6.1.0
platformdirs==4.3.6
pre_commit==4.0.1
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
radon==6.0.1
rich==13.9.4
six==1.17.0
sqlparse==0.5.3
stevedore==5.4.0
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
virtualenv==20.28.1
psycopg2==2.9.10
drf-spectacular==0.28.0
zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability
django-prometheus
confluent-kafka
boto3
